@@ -16,10 +16,10 @@
         }
     },
     "require": {
-        "php": ">=5.5.0"
+        "php": "^7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.0",
+        "phpunit/phpunit": "^6",
         "codeclimate/php-test-reporter": "dev-master"
     }
 }
