@@ -3,9 +3,9 @@
 namespace Lstr\Sprintf\Middleware;
 
 use Exception;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class DefaultTypeTest extends PHPUnit_Framework_TestCase
+class DefaultTypeTest extends TestCase
 {
     /**
      * @covers Lstr\Sprintf\Middleware\DefaultType::__construct
