@@ -2,9 +2,9 @@
 
 namespace Lstr\Sprintf\Middleware;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class MiddlewareAdapterTest extends PHPUnit_Framework_TestCase
+class MiddlewareAdapterTest extends TestCase
 {
     /**
      * @covers Lstr\Sprintf\Middleware\AbstractInvokable::__construct
